New viewmat for Jqt
Viewmat window contains several controls that allow
you to select the colormap, display slice if the input data
is more than 2D, and transpose the displayed slice or change
the way the values are scaled to colormap. In addition, next
to the visualized data, the currently used color scale and
ranges of covered values are displayed, as well as the range
of values of the visible slice, if scaling to the entire data
set is enabled.